+++ projects/demo/src/app/pages/badges/badges.ts
@@ -1,0 +1,9 @@
+import { ChangeDetectionStrategy, Component } from '@angular/core';
+
+@Component({
+  selector: 'demo-badges',
+  templateUrl: './badges.html',
+  styleUrl: './badges.scss',
+  changeDetection: ChangeDetectionStrategy.OnPush,
+})
+export class Badges {}