+++ projects/demo/src/app/pages/drawer/drawer.ts
@@ -1,0 +1,9 @@
+import { ChangeDetectionStrategy, Component } from '@angular/core';
+
+@Component({
+  selector: 'demo-drawer',
+  templateUrl: './drawer.html',
+  styleUrl: './drawer.scss',
+  changeDetection: ChangeDetectionStrategy.OnPush,
+})
+export class Drawer {}