+++ projects/demo/src/app/pages/inputs/inputs.ts
@@ -1,0 +1,9 @@
+import { ChangeDetectionStrategy, Component } from '@angular/core';
+
+@Component({
+  selector: 'demo-inputs',
+  templateUrl: './inputs.html',
+  styleUrl: './inputs.scss',
+  changeDetection: ChangeDetectionStrategy.OnPush,
+})
+export class Inputs {}