+++ projects/demo/src/app/pages/modal/modal.ts
@@ -1,0 +1,9 @@
+import { ChangeDetectionStrategy, Component } from '@angular/core';
+
+@Component({
+  selector: 'demo-modal',
+  templateUrl: './modal.html',
+  styleUrl: './modal.scss',
+  changeDetection: ChangeDetectionStrategy.OnPush,
+})
+export class Modal {}