+++ projects/demo/src/app/pages/progress/progress.ts
@@ -1,0 +1,9 @@
+import { ChangeDetectionStrategy, Component } from '@angular/core';
+
+@Component({
+  selector: 'demo-progress',
+  templateUrl: './progress.html',
+  styleUrl: './progress.scss',
+  changeDetection: ChangeDetectionStrategy.OnPush,
+})
+export class Progress {}