+++ projects/demo/src/app/pages/range/range.ts
@@ -1,0 +1,9 @@
+import { ChangeDetectionStrategy, Component } from '@angular/core';
+
+@Component({
+  selector: 'demo-range',
+  templateUrl: './range.html',
+  styleUrl: './range.scss',
+  changeDetection: ChangeDetectionStrategy.OnPush,
+})
+export class Range {}